LayoutTests/fast/dom/HTMLTemplateElement/cloneNode-expected.txt

Test that template contents are not cloned when the template element is cloned

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS template.content.childNodes.length is 1
PASS clone.content.childNodes.length is 0
